Updated README and the name of AWS KEY environmental variables; Make embedding service a separate folder.
