+++ data_ingestion/.env.template
@@ -1,13 +1,9 @@
 # Data Ingestion Environment Variables Template for Reg-Guru
 # Copy this file to .env and fill in the actual values
 
-# Data Path Configuration
-RAW_DATA_PATH='path/to/your/raw/data/folder'
-PROCESSED_DATA_PATH='path/to/your/processed/data/folder'
-
 # AWS S3 Configuration
-S3_ACCESS_KEY_ID='your-s3-access-key-id'
-S3_SECRET_ACCESS_KEY='your-s3-secret-access-key'
+AWS_ACCESS_KEY_ID='your-s3-access-key-id'
+AWS_SECRET_ACCESS_KEY='your-s3-secret-access-key'
 
 # AWS RDS PostgreSQL Configuration
 DB_HOST='your-database-host'
