+++ data_ingestion/src/common/S3Client.py
@@ -7,8 +7,8 @@ class S3Client:
     def __init__(self):
         self.client = boto3.client(
             's3',
-            aws_access_key_id=os.getenv("S3_ACCESS_KEY_ID"),
-            aws_secret_access_key=os.getenv("S3_SECRET_ACCESS_KEY")
+            aws_access_key_id=os.getenv("AWS_ACCESS_KEY_ID"),
+            aws_secret_access_key=os.getenv("AWS_SECRET_ACCESS_KEY")
         )
         return
