+++ data_ingestion/src/common/helper.py
@@ -55,8 +55,8 @@ def downloadPdftoS3(url, dest_key):
     # AWS S3 config
     s3_client = boto3.client(
         "s3",
-        aws_access_key_id=os.getenv("S3_ACCESS_KEY_ID"),
-        aws_secret_access_key=os.getenv("S3_SECRET_ACCESS_KEY"),
+        aws_access_key_id=os.getenv("AWS_ACCESS_KEY_ID"),
+        aws_secret_access_key=os.getenv("AWS_SECRET_ACCESS_KEY"),
         region_name=os.getenv("AWS_REGION")
     )
