+++ embedding_service/.env.template
@@ -1,0 +1,8 @@
+# Embedding service environmental variable template
+# ChromaDB
+CHROMADB_ROOT_DIR="chroma"
+# CORS Configuration (comma-separated list)
+CORS_ORIGINS="http://localhost:3000,http://127.0.0.1:3000"
+# Embedding Service
+EMBEDDING_MODEL="BAAI/bge-m3"
+EMBEDDING_SERVICE_URL="http://localhost:6000"
